+++ Kernel_23/doc/Kernel_23/CGAL/Kernel/global_functions.h
@@ -93,8 +93,9 @@ Kernel::FT approximate_angle(const CGAL::Point_3<Kernel>& p,
 /*!
 returns an approximation of the signed dihedral angle in the tetrahedron `pqrs` of edge `pq`.
 The sign is negative if `orientation(p,q,r,s)` is `CGAL::NEGATIVE` and positive otherwise.
-The angle is given in degrees.
+The angle is given in rad.
 \pre `p,q,r` and `p,q,s` are not collinear.
+\warning Before \cgal 4.14 the angle was given in degree.
 */
 template <typename Kernel>
 Kernel::FT approximate_dihedral_angle(const CGAL::Point_3<Kernel>& p,