+++ Installation/CHANGES.md
@@ -8,6 +8,8 @@ Release date: March 2019
 
 ### 2D and 3D Linear Geometry Kernel
 
+-   **Breaking change:** The function `approximate_dihedral_angle()` no longer returns
+    the angle in degree but in rad.
 -   Add concept `ComputeApproximate_kernel_3` to the `Kernel`. It has operators for computing
     the angle between two 3D vectors, as well as for three 3D points `p`, `q`, and `r`,
     representing the angles `p - q `, and `r - q`.